LayoutTests/platform/qt/http/tests/dom/window-open-about-blank-and-access-document-expected.txt

CONSOLE MESSAGE: line 24: Injected script running.

PASS newWindow.document is defined.
